Refactor: User account setting page and move FileSelector to EditView dir
